Reworked the Action node
The previous design only allowed for one AI at a time to use an ActionTree. What should really happen is the ActionTree decides, for each node, what the best action is. It should then delegate the processing of that action to a separate object (a reference in this case). That way, the nodes in the ActionTree can return to deciding the best action for a new employee