Made BaseCamera parentless
Now the BaseCamera doesn't parent itself to it's target. This has loads of benefits especially when the player is killed randomly. It also gives a lot of freedom on how the camera can follow the target. Also removed redundant code for making the camera glide to the player at the start. It was extremely restrictive and wonky