Simplified IK system, and added SqueezedCollider
Removed the need for ForwardTransform with some clever parenting. There is still a glitch when the character crouches, caused because IK nodes fully take over their bones from the AnimationTree. Also replaced the Body and Feet CollisionShape with just a Body shape which stretches between the feet and the head