Finally implemented a save feature
After a lot of trial and error, I finally came up with a system for saving nodes. It involves storing all variables in a node's script (if present), excluding the variables labelled in dont_save (if that is defined). Keep in mind the biggest limitation is that variables with Nodes and Resources cannot be saved. If a node defines _override_load, that Node will not be loaded by the Director, instead the information about the node's state is passed to _overeride_load. If _loading_state is defined, the node has the option of preparing some variables before the state is applied.