Implemented autosave and load on startup
The default scene is now a menu screen (with no ui) which will check if a save is present. Then it will redirect to either an empty scene to load a save, or the first level if there isn't a save. Nodes can add themselves to the ignore_state group so that their states are not stored (instead of adding a variable called ignore_state which required a script).